pyproject.toml
[tool.poetry]
name = "notion-songwriter"
version = "0.1.0"
description = "Converts Notion pages to songwriter-enhanced Markdown format"
authors = ["The <at.thealchemist@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
notion = "^0.0.28"
attrs = "^21.2.0"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
